# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Facturo API",
        "description": "API REST : authentification (Sanctum), clients, devis, factures, synthèse tableau de bord.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API locale — aligner avec APP_URL et L5_SWAGGER_CONST_HOST dans .env"
        }
    ],
    "paths": {
...
        "/api/invoices": {
            "get": {
                "tags": [
                    "Factures"
                ],
                "operationId": "2e947f49543378e8d1ac24fc3ebcdcc7",
                "responses": {
                    "200": {
                        "description": "Liste paginée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Factures"
                ],
                "operationId": "15892ddc0ad0649e0919d15345232fc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "subtotal",
                                    "tax_amount",
                                    "total"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "quote_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "draft",
                                            "sent",
                                            "paid",
                                            "overdue",
                                            "cancelled"
                                        ]
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "XOF",
                                        "nullable": true
                                    },
                                    "subtotal": {
                                        "type": "number"
                                    },
                                    "tax_amount": {
                                        "type": "number"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facture créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/invoices/{id}": {
            "get": {
                "tags": [
                    "Factures"
                ],
                "operationId": "5d25d392f44d2cf516114d871e72ecb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Factures"
                ],
                "operationId": "e5aa8abff3efb646c72e59724ecb13e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "quote_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "subtotal": {
                                        "type": "number"
                                    },
                                    "tax_amount": {
                                        "type": "number"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "paid_at": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Factures"
                ],
                "operationId": "410b8a2e4b097016c3dea391349a789b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum : header Authorization: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}